Laboratory/Projects/SplittingPrimer/BeepBeep-DataService/requirements.txt
pyjwt
-e git+https://github.com/giacomodeliberali/flakon.git#egg=flakon
flask_webtest
cryptography
sqlalchemy
flask_sqlalchemy
chaussette
flask_cors
requests_mock